Fix warning (Issue #301)

+++ mxml-search.c
@@ -117,7 +117,7 @@ mxmlFindPath(mxml_node_t *top,		// I - Top node
     if ((pathsep = strchr(path, '/')) == NULL)
       pathsep = path + strlen(path);
 
-    if (pathsep == path || (pathsep - path) >= sizeof(element))
+    if (pathsep == path || (size_t)(pathsep - path) >= sizeof(element))
       return (NULL);
 
     memcpy(element, path, pathsep - path);
